requirements.txt
alabaster==0.7.3
amqp==1.4.9
amqplib==1.0.2
anyjson==0.3.3
apsw==3.7.17.post1
attrs==16.0.0
Babel==2.3.4
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
BitTorrent==4.4.0
bottle==0.12.7
bzr==2.6.0
cairocffi==0.7.2
cffi==1.6.0
characteristic==14.3.0
chardet==2.3.0
Cheetah==2.4.4
CherryPy==3.6.0
click==6.6
ClientForm==0.2.10
Coffin==2.0.1
configobj==5.0.6
configparser==3.5.0
construct==2.5.5
crit==0.0.1
cryptography==1.3.4
cssselect==0.9.1
cssutils==1.0
cycler==0.9.0
decorator==3.4.2
Django==1.9.9
dnspython==1.15.0
docopt==0.6.2
docutils==0.12
ecdsa==0.13
enum34==1.1.3
Flask==0.11.1
futures==3.0.2
Genshi==0.7
gmixer==1.0
google-api-python-client==1.5.5
gunicorn==19.3.0
httplib2==0.9.2
idna==2.0
ipaddr==2.1.11
ipaddress==1.0.16
itsdangerous==0.24
java-gcj-compat==1.0.80
javapackages==4.6.0
Jinja2==2.8
keyring==9.3.1
kombu==3.0.33
libvirt-python==2.3.0
lxml==3.6.0
M2Crypto==0.24.0
MarkupSafe==0.23
matplotlib==1.5.1.post0
mechanize==0.2.5
mercurial==3.9.1
mod-python===3.5.0-none-none
mysql-utilities==1.5.6
nagstamon==1.0.1
ndg-httpsclient==0.4.0
netifaces==0.10.4
notify2==0.3
numpy==1.11.1
oauth2client==4.0.0
odfpy==0.9.6
osc===0.155git
oslosphinx==4.7.0
packaging==16.7
paramiko==2.0.2
pass-python-keyring==1.1
pbr==1.10.0
pdfrw==0.1
pexpect==4.2.1
pika==0.10.0
pika-pool==0.1.3
Pillow==3.1.1
pluggy==0.3.0
protobuf==2.6.1
psutil==4.3.1
ptyprocess==0.5
pulseaudio-dlna==0.5.2
py==1.4.31
pyalsa==1.0.29
pyalsaaudio==0.7
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyAudio==0.2.8
pybeam==0.3.2
pychecker==0.8.19
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pygame===1.9.1release
Pygments==2.1.3
pygobject==3.22.0
pyOpenSSL==16.0.0
PyPAM==0.5.0
pyparsing==2.0.7
pyPdf==1.13
pyserial==3.0.1
PySide==1.2.4
Python-Audio-Tools==2.19
python-dateutil==2.4.2
python-evtx==0.3.0
python-gflags==3.1.0
python-xlib==0.12
pytz==2016.4
pyudev==0.20.0
PyXB==1.2.4
PyYAML==3.12
reportlab==3.2.0
repoze.sphinx.autointerface==0.7.1
requests==2.11.1
rsa==3.4.2
## FIXME: could not find svn URL in dependency_links for this package:
rst2pdf===0.93.dev-r0
scapy==2.2.0
scipy==0.17.0
SecretStorage==2.3.1
service-identity==16.0.0
setproctitle==1.1.8
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.6
sphinx-rtd-theme==0.1.8
svg2rlg==0.3
tornado==4.3
tox==2.3.1
Twisted==16.4.0
uritemplate==3.0.0
urlgrabber==3.9.1
virtualenv==15.0.0
Werkzeug==0.11.11
wxPython-common==3.0.2.0
zenmap==7.31
zeroconf==0.17.6
zope.interface==4.1.2
